+++ packages/client/src/pages/forum/styles/create.pcss
@@ -1,0 +1,5 @@
+.forum-create-actions {
+  display: flex;
+  gap: 30px;
+  width: 800px;
+}
